\name{athNet}
\docType{data}
\alias{athNet}
\title{Example network data set for CoReg}
\description{
  This is the example network data set for CoReg package. This network is a merged network generated by two publications for Arabidopsis. (See references)
}
\usage{data(athNet)}
\format{An igraph object containing 412 nodes and 1490 edges.}
\references{
  M. Taylor-Teeples et al. (2015) An Arabidopsis gene regulatory network for secondary cell wall synthesis. Nature 2015 517:571-575.
  
  Erin E. Sparks et al. (2016) Establishment of Expression in the SHORTROOT-SCARECROW Transcriptional Cascade through Opposing Activities of Both Activators and Repressors. Developmental Cell 2016 39:585-596.
}
\keyword{Arabidopsis network}