\name{CoReg-package}
\alias{CoReg-package}
\alias{CoReg}
\docType{package}
\title{CoReg co-regulation module finding}
\description{
    CoReg identifies the co-regulation module(s) from the regulatory networks
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CoReg}
\packageIndices{CoReg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Qi Song

Maintainer: Qi Song <alexsong@vt.edu>
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{ 
}
