\name{CoReg-package}
\alias{CoReg-package}
\docType{package}
\title{CoReg co-regulation module finding}
\description{
    CoReg identifies the co-regulation module(s) from the regulatory networks
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CoReg}
\packageIndices{CoReg}
CoReg identifies the co-regulatory module(s) from the regulatory networks. 
}
\author{
Qi Song

Maintainer: Qi Song <alexsong@vt.edu>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{ 
}
